% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barPlotAncest.R
\name{barPlotAncest}
\alias{barPlotAncest}
\title{Create a Barplot of Ancestry Proportion Estimates for
Every Sample and a Given Chromosome or Chromosomal Region}
\usage{
barPlotAncest(set, order = TRUE, title = "", xlab = "Sample",
  ylab = "Ancestry Proportion", ...)
}
\arguments{
\item{set}{A \code{data.frame} with columns of the proportion
ancestry for a given chromosome or chromosomal region, and one
row per sample (bar).}

\item{order}{A \code{logical} argument determining whether the
samples should be ordered in increasing proportion of the first
ancestry. Default is TRUE.}

\item{title}{A character string containing the title of the plot.
Default is "", a blank title.}

\item{xlab}{A character vector with the label for the x-axis on
the plot. Default is \code{Sample}.}

\item{ylab}{A character vector holding the label for the y-axis
on the plot. Default is \code{Ancestry Proportion}.}

\item{...}{Further arguments to be passed to the plotting methods,
such as graphical parameters.}
}
\value{
Creates a plot.
}
\description{
Plots ancestry proportion estimates for each sample
}
\details{
Creates a barplot of ancestry proportions for each sample
for a given chromosome or chromosomal region.
}
\examples{
data(ancestries)
chr1 <- ancestries[,c("Euro_1","Afr_1","Asian_1")]
#barPlotAncest(chr1,title="Chr 1 Ancestry Proportions")
}
\author{
Caitlin McHugh \email{mchughc@uw.edu}
}

