% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_combP.R
\name{calc_combP}
\alias{calc_combP}
\title{Calculate the Combined CAnD Test Statistic P-value}
\usage{
calc_combP(chrAncest)
}
\arguments{
\item{chrAncest}{A data.frame holding the ancestral proportions;
each row corresponds to a sample and each column corresponds to a
chromosomal/chromosomal segment ancestry proportion.
Note: only include the proportions for one ancestral population at a time.}
}
\value{
A \code{vector} of length two where `statistic' is the
combined CAnD statistic and `pvalue' is it's corresponding p-value,
where the combined statistic is combined over all
chromosomes/chromosomal segments included in \code{chrAncest}.
}
\description{
Calculate the combined CAnD test statistic p-value on a set of
ancestry proportions estimated for a particular ancestral
subpopulation of interest
}
\examples{
data(ancestries)
euroCols <- grep("Euro",colnames(ancestries))
euro <- ancestries[,euroCols]
res <- calc_combP(euro)
res
}
\author{
Caitlin McHugh
\email{mchughc@uw.edu}
}
\references{
McHugh, C., Brown, L., Thornton, T.
Detecting heterogeneity in population structure across the genome in admixed populations.
Genetics, 2016.
}

