% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDiffMatrices.R
\name{getDiffMatrices}
\alias{getDiffMatrices}
\title{Calculate the Mean Ancestry Proportion Excluding Each
Chromosome/Chromosomal Segment in Turn}
\usage{
getDiffMatrices(chrAncest, diff = TRUE)
}
\arguments{
\item{chrAncest}{A data.frame holding the ancestral proportions;
each row corresponds to a sample and each column corresponds to a
chromosomal/chromosomal segment ancestry proportion.
Note: only include the proportions for one ancestral population at a time.}

\item{diff}{A logical argment indicating whether the difference
between the pooled mean and the chromosomal mean should be returned,
or whether simply the pooled mean should be returned.}
}
\value{
A matrix of chromosomal ancestry differences.
}
\description{
A helper function to calculate the mean ancestry proportion for a
given subpopulation, excluding each chromosome/chromosomal segment
in turn.
}
\details{
This function calculates the mean ancestry proportion of a given
subpopulation excluding each chromosome in turn.
}
\author{
Caitlin McHugh
\email{mchughc@uw.edu}
}

