% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nonParam_CAnD.R
\name{nonParam_CAnD}
\alias{nonParam_CAnD}
\title{Perform the Non-Parametric CAnD Test}
\usage{
nonParam_CAnD(chrAncest, bonfCorr = TRUE)
}
\arguments{
\item{chrAncest}{A data.frame holding the ancestral proportions;
each row corresponds to a sample and each column corresponds to a
chromosomal/chromosomal segment ancestry proportion.
Note: only include the proportions for one ancestral population at a time.}

\item{bonfCorr}{A logical argument indicating whether the p-value
should be corrected for multiple testing using Bonferroni
correction. The default is \code{TRUE}.}
}
\value{
A \code{CAnDResult} object holding the p-value for each
chromosome/chromosomal segment, the overall CAnD p-value, the
CAnD statistic, and whether the Bonferroni multiple testing
correction was used.
}
\description{
Perform the non-parametric CAnD test on a set of ancestry proportions
estimated for a particular ancestral subpopulation of interest.
}
\examples{
data(ancestries)
afrCols <- grep("Afr",colnames(ancestries))
afr <- ancestries[,afrCols]
res <- nonParam_CAnD(afr)
print(res)
}
\author{
Caitlin McHugh
\email{mchughc@uw.edu}
}
\references{
McHugh, C., Brown, L., Thornton, T.
Detecting heterogeneity in population structure across chromosomes in admixed populations.
Manuscript in Preparation.
}

