% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPvals.R
\name{plotPvals}
\alias{plotPvals}
\title{Create a Plot of P-Values for Each Chromosome or
Chromosomal Region}
\usage{
plotPvals(set, title = "", xlab = "Chromosome", ylab = "-log10(PValue)",
  ...)
}
\arguments{
\item{set}{An object of class \code{CAnDResult}.}

\item{title}{A character string containing the title of the plot.
Default is "", a blank title.}

\item{xlab}{A character vector with the label for the x-axis on
the plot. Default is \code{Chromosome}.}

\item{ylab}{A character vector holding the label for the y-axis
on the plot. Default is \code{-log10(Bonferroni PValue)}
or \code{-log10(PValue)},
depending on whether Bonferroni correction was used.}

\item{...}{Further arguments to be passed to the plotting methods,
such as graphical parameters.}
}
\value{
Creates a plot.
}
\description{
Plots CAnD p-values for each chromosome/chromosomal region
}
\details{
Creates a plot of all p-values for each chromosome or
chromosomal region.
}
\examples{
data(ancestries)
euroEsts <- ancestries[,c(seq(from=2,to=24))]
res <- CAnD(euroEsts)
plotPvals(res,main="CAnD P-Values")
}
\author{
Caitlin McHugh \email{mchughc@uw.edu}
}

